put sanity checks in their own module
